snu2/requirements.txt
openai>=1.0.0
streamlit>=1.28.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
